requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.3.4
    # via django
certifi==2021.5.30
    # via
    #   elastic-transport
    #   elasticsearch
click==8.0.1
    # via flask
cycler==0.10.0
    # via matplotlib
django==3.2.4
    # via -r requirements.in
eland==7.13.0b1
    # via -r requirements.in
elastic-enterprise-search==7.13.0
    # via -r requirements.in
elastic-transport==7.13.0
    # via elastic-enterprise-search
elasticsearch-dsl==7.3.0
    # via -r requirements.in
elasticsearch==7.13.2
    # via
    #   -r requirements.in
    #   eland
    #   elasticsearch-dsl
flask==2.0.1
    # via -r requirements.in
itsdangerous==2.0.1
    # via flask
jinja2==3.0.1
    # via flask
kiwisolver==1.3.1
    # via matplotlib
markupsafe==2.0.1
    # via jinja2
matplotlib==3.4.2
    # via eland
numpy==1.21.0
    # via
    #   eland
    #   matplotlib
    #   pandas
pandas==1.2.5
    # via eland
pillow==8.2.0
    # via matplotlib
pyjwt==2.1.0
    # via elastic-enterprise-search
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.1
    # via
    #   elastic-enterprise-search
    #   elasticsearch-dsl
    #   matplotlib
    #   pandas
pytz==2021.1
    # via
    #   django
    #   pandas
six==1.16.0
    # via
    #   elastic-enterprise-search
    #   elastic-transport
    #   elasticsearch-dsl
    #   python-dateutil
sqlparse==0.4.1
    # via django
urllib3==1.26.5
    # via
    #   elastic-transport
    #   elasticsearch
werkzeug==2.0.1
    # via flask
